==> notebooks/2_models/04_data_cleaning.ipynb <==
# Introduction

This notebook is dedicated to preparing and organizing the RAG-12000 dataset for downstream tasks such as training, hyperparameter tuning, and evaluation.

We begin by loading and cleaning the dataset, removing null entries and duplicates, and filtering meta-questions. Then we split the dataset into three main parts:

- **Fine-tuning Set** (70%)
- **Hyperparameter Tuning Set** (15%)
- **Test Set** (15%)

#### 1. Dataset Loading and cheking for missing entries/duplicates

!git clone https://huggingface.co/datasets/neural-bridge/rag-dataset-12000

import pandas as pd

test_data=pd.read_parquet("/kaggle/working/rag-dataset-12000/data/test-00000-of-00001-af2a9f454ad1b8a3.parquet")
train_data=pd.read_parquet("/kaggle/working/rag-dataset-12000/data/train-00000-of-00001-9df3a936e1f63191.parquet")
all_data=pd.concat([train_data,test_data])

#none?
print(f" Context has {all_data['context'].isna().sum()}  none values\n Question has {all_data['question'].isna().sum()} none values\n Answer has {all_data['answer'].isna().sum()} none values")

cleaned_data=all_data.dropna()

#duplicates
print(f" Context has {len(cleaned_data['context'].values)-len(cleaned_data['context'].unique())}  duplicates \n Question has {len(cleaned_data['question'].values)-len(cleaned_data['question'].unique())} duplicates\n Answer has {len(cleaned_data['answer'].values)-len(cleaned_data['answer'].unique())} duplicates")

question_count=cleaned_data.groupby("question").size()
question_count[question_count>1]

#search for potential questions related to context
context_queries=cleaned_data[cleaned_data["question"].str.contains('context', case=False)]["question"]
context_queries.shape
context_queries.head(100).values

#### Around 8% of the dataset was removed because it contained meta-questions mentioning "context", which are not representative of natural user inputs.


final_data=cleaned_data[~cleaned_data["question"].str.contains("context",case=False)]

final_data_no_duplicates=final_data.drop_duplicates("question")
final_data_no_duplicates

#### General data cleaning

import re
from bs4 import BeautifulSoup

def clean_text(text):
    if not isinstance(text, str):
        return ""
    text = text.lower()
    text = BeautifulSoup(text, "html.parser").get_text()
    text = re.sub(r"\s+", " ", text).strip()
    return text
final_data_no_duplicates["context"]=final_data_no_duplicates["context"].apply(lambda x : clean_text(x))

#### Data splitting & save

from sklearn.model_selection import train_test_split
finetuning_data, tmp_data = train_test_split(final_data_no_duplicates, test_size=0.3, random_state=42)
hyperparameter_tuning_data, test_data=train_test_split(tmp_data, test_size=0.5, random_state=42)

finetuning_data.to_parquet("/kaggle/working/finetuning_data.parquet")
hyperparameter_tuning_data.to_parquet("/kaggle/working/hyperparameter_tuning_data.parquet")
test_data.to_parquet("/kaggle/working/test_data.parquet")
