==> notebooks/2_models/06_semantic_search.ipynb <==
!pip install hnswlib



!ls /kaggle/input/nlp-resources/test_data.parquet


import pandas as pd


all_data=pd.read_parquet("/kaggle/input/nlp-resources/cleaned_data.parquet")
labels = pd.read_parquet("/kaggle/input/labeled/df_supervised.parquet")


lab = labels['cluster_label'].tolist()

Let's load a sentence transformer to compute the embeddings and a cross model encoder for reranking

from sentence_transformers import SentenceTransformer, CrossEncoder, util

semb_model = SentenceTransformer('multi-qa-MiniLM-L6-cos-v1')
xenc_model = CrossEncoder('cross-encoder/ms-marco-MiniLM-L-6-v2') #they work better but they need the compute the embedding every time we have a new query

Let's now compute the context embedding we will use for answering

import torch
contexts = all_data['context'].tolist()
corpus_embeddings = semb_model.encode(contexts, convert_to_tensor=True, show_progress_bar=True)
#corpus_embeddings = torch.load('//kaggle/input/corpusembeddings/corpus_embeddings.pt')


Let's now use some random sentence, one for each cluster previously identified: we will use cosine similarity to see what is the most similar context and if it belongs to the same category

queries = [
    "She uploaded the latest app update to the cloud-based platform without a hitch.",           # Digital/Online Services
    "The director revealed the teaser trailer at the midnight screening.",                      # Film & Television
    "The assembly line hummed as robotic arms welded the steel frame with precision.",           # Industrial & Machinery
    "He laced up his cleats just before kick-off under the stadium lights.",                    # Sports & Athletics
    "She diversified her portfolio by investing in emerging market bonds last quarter.",        # Finance & Investment
    "They sold out tickets to the championship gaming tournament in under an hour.",             # Gaming & Events
    "Guests strolled through the resort’s marble lobby, overlooking the sun-drenched terrace.",  # Hospitality & Real Estate
    "He spent the afternoon lost in a novel about existentialism and the human condition.",      # Literature & Philosophy
    "The clinic introduced a cutting-edge diagnostic tool to improve patient outcomes.",         # Healthcare & Medicine
    "She curated a capsule wardrobe to streamline her daily style routine.",                     # Lifestyle & Personal Interests
    "Lawmakers convened to draft the new regulations on data privacy and trade compliance.",    # Legal & Government
    "He savored the flaky croissant while reading the recipe for boulangerie perfection.",      # Food & Culinary
    "The candidate outlined her vision for education reform during the televised debate.",       # Politics & Public Affairs
    "Volunteers conducted a beach cleanup to protect the fragile coastal ecosystem.",           # Environment & Ecology
    "She blended acrylics to capture the vibrant hues of the urban skyline at dusk.",            # Art & Design
    "Researchers presented their groundbreaking findings on renewable energy sources.",         # Education & Research
    "The DJ mixed vintage vinyl records to curate the perfect summer soundtrack.",              # Music & Audio
    "He practiced mindfulness meditation to maintain balance amid his hectic schedule."         # Wellness & Personal Care
]


Let's compute the embeddings of these sentences, too.

queries_embeddings = semb_model.encode(queries, convert_to_tensor=True,show_progress_bar=True )

Let's find the most similar contexts with cosine similarity and see what category they belong to and if they are actually relevant.

We can also use the built-in function for semantic search

query_embeddings =  semb_model.encode(queries, convert_to_tensor=True)

util.semantic_search(query_embeddings, corpus_embeddings, score_function=util.cos_sim)

We can also speed up this process by pre-normalising the embeddings and compute directly the dot-product instead of cosine similarity.

normalised_docs_embeddings = util.normalize_embeddings(semb_model.encode(contexts, convert_to_tensor=True))
normalised_query_embeddings = util.normalize_embeddings(semb_model.encode(queries, convert_to_tensor=True))
#norm_docs_embeddings = torch.load('/kaggle/input/normdocsembedding/norm_docs_embeddings.pt')
#norm_query_embeddings = torch.load('/kaggle/input/normqueryemb/norm_query_embeddings.pt')

hits = util.semantic_search(normalised_query_embeddings, normalised_docs_embeddings, score_function=util.dot_score)
hits

Since it took a lot of time, let's save the embedding we previously computed

torch.save(corpus_embeddings, 'corpus_embeddings.pt')
torch.save(normalised_docs_embeddings, 'norm_docs_embeddings.pt')
torch.save(normalised_query_embeddings, 'norm_query_embeddings.pt')

In order to load them the next time i have to do this

'''corpus_embeddings = torch.load('corpus_embeddings.pt')
norm_docs_embeddings = torch.load('normalised_docs_embeddings.pt')
norm_query_embeddings = torch.load('normalised_query_embeddings.pt')'''

Let's now try and do the same thing with the cross encoder. Basically this model will be used to compute the similarity instead of using cosine similarity.

import numpy as np

import textwrap  

MAX_CHARS = 150
TOP_K = 5

for query in queries:
    
    xenc_model_inputs = [[query, context] for context in contexts]
    # compute similaruty
    scores = xenc_model.predict(xenc_model_inputs)

    # i take highest TOP_K 
    top_idxs = np.argsort(-scores)[:TOP_K]

    print(f"Query: \"{query}\"")
    print("Top 5 most similar contexts:")
    print("---------------------------------------")
    for idx in top_idxs:
        full_ctx = contexts[idx]
        # hortening contex
        short_ctx = full_ctx[:MAX_CHARS] + ("..." if len(full_ctx) > MAX_CHARS else "")
       

        print(f"Score: {scores[idx]:.4f} – \"{short_ctx}\", {lab[idx]}")
    print("\n")

Very nice: as we can see in many cases the labels match!

When we use an embedding model with cosine similarity, we can pre-compute the embeddings in our data set and index them to speed-up the search. There are techniques for Approximate Nearest Neighbor (ANN), which use clustering to index the embedding space and speed-up the search process. Let's do indexing with HNSWLIB

import hnswlib

index = hnswlib.Index(space='cosine', dim=corpus_embeddings.size(1))

Now we can index our data. The index we compute can be saved and loaded, so we can check if it is already availabel and load it (this will save time)

import os

# Define hnswlib index path
index_path = "./hnswlib.index"

# Load index if available
if os.path.exists(index_path):
    print("Loading index...")
    index.load_index(index_path)
# Else index data collection
else:
    # Initialise the index
    print("Start creating HNSWLIB index")
    index.init_index(max_elements=corpus_embeddings.size(0), ef_construction=400, M=64)
    #  Compute the HNSWLIB index (it may take a while)
    index.add_items(corpus_embeddings.cpu(), list(range(len(corpus_embeddings))))
    # Save the index to a file for future loading
    print("Saving index to:", index_path)
    index.save_index(index_path)

Let's see if it's actually faster by measuring the running time

from datetime import datetime

# Search using index
t_start = datetime.now()
_ = index.knn_query(queries_embeddings[0].cpu(), k=128)
t_stop = datetime.now()
print(f"Search time with index: {t_stop - t_start}")

# Search without index
t_start = datetime.now()
_ = util.semantic_search(queries_embeddings, corpus_embeddings, score_function=util.cos_sim, top_k=128)
t_stop = datetime.now()
print(f"Search time without index: {t_stop - t_start}")

Reranking
Since the cross encoder gives better results but is slower than cosine similarity, we can take advantage of both: We can do a first search with bi-encoder models and then re-rank the top- k  results with a cross-encoder. We call this approach retrieve and re-rank.

Let's define a new random query (we could have used one of those generated before)

query = "Who is the president of the United States?"
query_embedding = semb_model.encode(query, convert_to_tensor=True)

Search using ANN index

corpus_ids, distances = index.knn_query(query_embedding.cpu(), k=3)
scores = 1 - distances

print("Cosine similarity model search results")
print(f"Query: \"{query}\"")
print("---------------------------------------")
for idx, score in zip(corpus_ids[0], scores[0]):
    print(f"Score: {score:.4f}\nDocument: \"{contexts[idx]}\"\n\n")

Let's add the re-ranking

corpus_ids, _ = index.knn_query(query_embedding.cpu(), k=128)

model_inputs = [(query, contexts[idx]) for idx in corpus_ids[0]]
cross_scores = xenc_model.predict(model_inputs)

print("Cross-encoder model re-ranking results")
print(f"Query: \"{query}\"")
print("---------------------------------------")
for idx in np.argsort(-cross_scores)[:3]:
    print(f"Score: {cross_scores[idx]:.4f}\nDocument: \"{contexts[corpus_ids[0][idx]]}\"\n\n")

Finally, let's see if our dataset is big enough to give correct answers to generic questions just by finding the most similar question in the dataset and giving its associated answer. It should work kind of like a chatbot

Let's first compute the answers' embeddings

answers = all_data['answer'].tolist()
answers_embeddings = semb_model.encode(answers, convert_to_tensor=True, show_progress_bar=True)

questions = all_data['question'].tolist()
questions_embeddings = semb_model.encode(questions, convert_to_tensor=True, show_progress_bar=True)

# Create empty index
hnswlib_index = hnswlib.Index(space='cosine', dim=questions_embeddings.size(1))

# Define hnswlib index path
index_path = "./emp_dialogue_hnswlib.index"

# Load index if available
if os.path.exists(index_path):
    print("Loading index...")
    hnswlib_index.load_index(index_path)
# Else index data collection
else:
    # Initialise the index
    print("Start creating HNSWLIB index")
    hnswlib_index.init_index(max_elements=questions_embeddings.size(0), ef_construction=400, M=64)
    #  Compute the HNSWLIB index (it may take a while)
    hnswlib_index.add_items(questions_embeddings.cpu(), list(range(len(questions_embeddings))))
    # Save the index to a file for future loading
    print("Saving index to:", index_path)
    hnswlib_index.save_index(index_path)

import numpy as np
def get_response(
    message: str,
    questions_embeddings,     # array/tensor of questions embeddings
    answers: list[str],       # answers list
    index,                    # hnswlib.Index with questions_embeddings
    re_ranking_model=None,    #  cross-encoder
    top_k: int = 32
) -> str:
   
    message_emb = semb_model.encode(message, convert_to_tensor=True).cpu().numpy()

   
    corpus_ids, _ = index.knn_query(message_emb, k=top_k)

    
    candidate_idxs = corpus_ids[0]
    if re_ranking_model is not None:
        model_inputs = [
            (message, answers[i])
            for i in candidate_idxs
        ]
        cross_scores = re_ranking_model.predict(model_inputs) 
        best_pos = np.argmax(cross_scores)
        best_idx = candidate_idxs[best_pos]
    else:
       
        best_idx = candidate_idxs[0]
    return answers[best_idx]

chatbot_response = get_response(
    "who is einstein?", 
    questions_embeddings,                
    answers,                            
    hnswlib_index,                        
    re_ranking_model=xenc_model,          
    top_k=32                              
)

print("Chatbot says:", chatbot_response)


As we could imagine, not the best result even thought if we are lucky the responses actully make some sense. Of course this is because the dataset we have is not that big
