==> notebooks/2_models/07_t5.ipynb <==
## Introduction
This notebook evaluates the T5-small model's performance on question answering tasks 
We analyze:
- Zero-shot capabilities
- Few-shot learning
- Context vs no-context comparisons
- Fine-tuned performance 

## 1. Dataset Loading and Utility Functions

import pandas as pd
finetuning_data=pd.read_parquet("/kaggle/input/nlp-resources/tuning_data.parquet")
hyperparameter_tuning_data=pd.read_parquet("/kaggle/input/nlp-resources/hyperparameter_tuning_data.parquet")
test_data=pd.read_parquet("/kaggle/input/nlp-resources/test_data.parquet")
contexts=pd.concat([test_data["context"],hyperparameter_tuning_data["context"],finetuning_data["context"]]).to_list()


import torch
import pandas as pd
from transformers import T5Tokenizer
from datasets import load_dataset
from datasets import Dataset
from transformers import T5ForConditionalGeneration, Trainer, TrainingArguments, DataCollatorForSeq2Seq
import re
import string
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score
from transformers import EarlyStoppingCallback
from transformers import TrainerCallback
from functools import reduce
import nltk
from nltk.translate.bleu_score import sentence_bleu, SmoothingFunction
from sklearn.metrics.pairwise import cosine_similarity
import string
import re
from collections import Counter
import numpy as np
nltk.download('punkt')

def preprocess(example,tokenizer,max_length_input=512,max_length_labels=64,no_context=False):
    input_text = f"question: {example['question']}" 
    if (not no_context):
        input_text+=f" context: {example['context']}"
    target_text = example["answer"]
    inputs = tokenizer(input_text, max_length=max_length_input, truncation=True, padding="max_length")
    labels = tokenizer(target_text, max_length=max_length_labels, truncation=True, padding="max_length")["input_ids"] 
    inputs["labels"] =labels
    return inputs

def generate_answer(model, tokenizer, question, device, max_input_length=512, max_output_length=64, context=None):
    model.eval()
    input_text = f"question: {question}"
    if(context):
        input_text+= f" context: {context}"
    inputs = tokenizer.encode(input_text, return_tensors="pt", max_length=max_input_length, truncation=True)
    inputs = inputs.to(device)

    with torch.no_grad():
        outputs = model.generate(
            inputs,
            max_length=max_output_length,
            num_beams=4,
            early_stopping=True,
            do_sample=True,
            top_k=50,
            top_p=0.95,
            temperature=0.9,
            repetition_penalty=1.2,
            no_repeat_ngram_size=3
        
        )

    return tokenizer.decode(outputs[0], skip_special_tokens=True)

def normalize_answer(s):
    def remove_articles(text):
        return re.sub(r'\b(a|an|the)\b', ' ', text)
    def remove_punc(text):
        return ''.join(ch for ch in text if ch not in set(string.punctuation))
    def white_space_fix(text):
        return ' '.join(text.split())
    def lower(text):
        return text.lower()

    return white_space_fix(remove_articles(remove_punc(lower(s))))

def exact_match_score(prediction, ground_truth):
    return int(normalize_answer(prediction) == normalize_answer(ground_truth))

def f1_score(prediction, ground_truth):
    pred_tokens = normalize_answer(prediction).split()
    gt_tokens = normalize_answer(ground_truth).split()
    common = Counter(pred_tokens) & Counter(gt_tokens)
    num_same = sum(common.values())
    if num_same == 0:
        return 0
    precision = num_same / len(pred_tokens)
    recall = num_same / len(gt_tokens)
    return 2 * precision * recall / (precision + recall)
    
def bleu_score(prediction, ground_truth):
    smoothie = SmoothingFunction().method4  # migliora BLEU su frasi corte
    pred_tokens = nltk.word_tokenize(prediction.lower())
    gt_tokens = nltk.word_tokenize(ground_truth.lower())
    return sentence_bleu([gt_tokens], pred_tokens, smoothing_function=smoothie)
    
def get_top_k_contexts(context_embeddings,model_embed,query,k=3):
    query_emb=model_embed.encode([query], convert_to_numpy=True)
    similarity=cosine_similarity(query_emb,context_embeddings)[0]
    #print(similarity)
    top_k_idx = similarity.argsort()[-k:]
    return [contexts[i] for i in top_k_idx]
    
def eval_answers(model,tokenizer,test_data=test_data,max_input=512,max_output=64, rag=False, no_context=False):
    exact_matches = []
    f1_scores = []
    bleu_scores=[]
    for idx, row in test_data.iterrows():
        context= None if no_context else row['context']
        pred_answer = generate_answer(model, tokenizer, row['question'],max_input_length=max_input,max_output_length=max_output, device=device,context=context )
        em = exact_match_score(pred_answer, row['answer'])
        f1 = f1_score(pred_answer, row['answer'])
        bleu = bleu_score(pred_answer, row['answer'])
        exact_matches.append(em)
        f1_scores.append(f1)
        bleu_scores.append(bleu)
    return sum(bleu_scores) / len(bleu_scores) * 100,sum(exact_matches) / len(exact_matches) * 100, sum(f1_scores) / len(f1_scores) * 100
def print_scores(avg_bleu=-1, avg_em=-1, avg_f1=-1,model_name="",scores_path="/kaggle/input/nlp-resources/scores_t5_rag.parquet",use_backup=False):
    
    if(use_backup):
        scores=pd.read_parquet(scores_path)
        scores_array=scores[scores["model_name"]==model_name][["avg_bleu","avg_em","avg_f1"]].values[0]
    else:
        scores_array=np.array([avg_bleu, avg_em, avg_f1])
    print(f"Scores of {model_name} model :")
    print(f"Average BLEU score: {scores_array[0]:.2f}")
    print(f"Exact Match: {scores_array[1]:.2f}%")
    print(f"F1 Score: {scores_array[2]:.2f}%")
device=torch.device("cuda" if torch.cuda.is_available() else "cpu")

## 2. Evaluation Metrics

We use three key metrics:

| **Metric**       | **Description**         | **Importance**                |
|------------------|-------------------------|-------------------------------|
| Exact Match (EM) | Strict string matching  | Measures precision            |
| F1 Score         | Token-level overlap     | Balances precision/recall     |
| BLEU Score       | N-gram similarity       | Evaluates fluency             |


## 3. Zero-Shot Evaluation

from transformers import T5ForConditionalGeneration, Trainer, TrainingArguments, DataCollatorForSeq2Seq

tokenizer_T5 = T5Tokenizer.from_pretrained("t5-small")
model_T5_pretrained = T5ForConditionalGeneration.from_pretrained("t5-small")
device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
model_T5_pretrained.to(device)
model_T5_pretrained.eval()

#### Zero shot sample :

sample_zero_shot=test_data.sample(n=1,random_state=42)
print(f"Sampled exapmle:\n#Question: {sample_zero_shot['question'].values[0]}\n#Context:{sample_zero_shot['context'].values[0]}")

#### Zero shot generated output :

zero_shot_answer=generate_answer(model_T5_pretrained, tokenizer_T5, question=sample_zero_shot['question'].values[0], device=device, max_input_length=512, max_output_length=128, context=sample_zero_shot['context'].values[0])
print(f"Genereated answer:{zero_shot_answer}\n\nDataset Answer:{sample_zero_shot['answer'].values[0]}")
print(f"\nComparison with dataset answer:\nF1:{f1_score(zero_shot_answer, sample_zero_shot['answer'].values[0]):.2f}",f"BLEU:{bleu_score(zero_shot_answer, sample_zero_shot['answer'].values[0]):.2f}",f"EM:{exact_match_score(zero_shot_answer, sample_zero_shot['answer'].values[0]):.2f}")

## 4. Few-Shot Prompt (1-Shot)

We manually construct a prompt that includes one example (1-shot learning). This tests the model's ability to learn from a single example.


test_data["len"]=test_data["context"].apply(len)
samples=test_data.sort_values(by="len").head(2)
shots=samples.head(1)
test_sample=samples.iloc[-1]
promt=""
index=1
for i,row in shots.iterrows():
    example=f"Example {index} question: {row['question']} context: {row['context']} answer: {row['answer']}\n"
    promt+=example
    index+=1
promt+=f"question: {test_sample['question']} context: {test_sample['context']} answer:"
print(promt)

model_T5_pretrained.eval()
inputs = tokenizer_T5(promt, return_tensors="pt", max_length=2048, truncation=True)
inputs = {k: v.to(device) for k, v in inputs.items()}
with torch.no_grad():
    outputs = model_T5_pretrained.generate(
        **inputs,
        max_length=128,
        num_beams=4,
        early_stopping=True
    )
gen_answer_1_shot=tokenizer_T5.decode(outputs[0], skip_special_tokens=True).strip()
gen_answer_zero_shot=generate_answer(model_T5_pretrained, tokenizer_T5, question=test_sample['question'], device=device, max_input_length=512, max_output_length=64, context=test_sample['context'])

### Let's compare the generated answer with the real one

print(f"Question: {test_sample['question']}\nGenereated one shot answer:{gen_answer_1_shot}\nGenereated zero shot answer:{gen_answer_zero_shot}\nDataset Answer:{test_sample['answer']}")
print(f"Comparison with real answer \nF1:{f1_score(gen_answer_1_shot,test_sample['answer']):.2f} BLEU:{bleu_score(gen_answer_1_shot,test_sample['answer']):.2f} EM:{exact_match_score(gen_answer_1_shot,test_sample['answer']):.2f}")

#### Aggregate Performance of zero shot calculated on test data (15% all data)

#avg_bleu, avg_em, avg_f1=eval_answers(model_T5_pretrained, tokenizer_T5,test_data,max_input=512,max_output=64)
print_scores(model_name="t5_zero_shot",use_backup=True)

## 5. Zero-Shot Evaluation Without Context

This step evaluates the model in a context-free setting to understand how much background information contributes to performance.


#avg_bleu, avg_em, avg_f1=eval_answers(model_T5_pretrained,tokenizer_T5,test_data,no_context=True)
print_scores(model_name="t5_zero_shot_no_context",use_backup=True)

## 6. Model Fine-Tuning

We fine-tune the pre-trained T5-small model on our QA dataset. This helps the model adapt to the specific language and structure of the task.

from transformers import T5ForConditionalGeneration, Trainer, TrainingArguments, DataCollatorForSeq2Seq

tokenizer_T5_tuned = T5Tokenizer.from_pretrained("t5-small")
model_T5_tuned = T5ForConditionalGeneration.from_pretrained("t5-small")
model_T5_tuned.to(device)
train,val = train_test_split(finetuning_data, test_size=0.2, random_state=42)
train_dataset = Dataset.from_pandas(train)
val_dataset = Dataset.from_pandas(val)
tokenizer=tokenizer_T5_tuned
train_dataset = train_dataset.map(preprocess,batched=False,fn_kwargs={
        "tokenizer":tokenizer_T5_tuned,
        "max_length_input": 512,
        "max_length_labels": 64
    })
val_dataset = val_dataset.map(preprocess,batched=False,fn_kwargs={
        "tokenizer":tokenizer_T5_tuned,
        "max_length_input": 512,
        "max_length_labels": 64
    })
training_args = TrainingArguments(
    output_dir="/kaggle/working",
    do_train=True,
    do_eval=True,
    learning_rate=0.000472,
    per_device_train_batch_size=4,
    per_device_eval_batch_size=4,
    num_train_epochs=6,
    weight_decay=0.041494,
    logging_dir="./logs",
    logging_steps=500,
    save_steps=1000,     
    eval_steps=1000,      
    save_total_limit=2,
    report_to="none",
)
trainer = Trainer(
    model=model_T5_tuned,
    args=training_args,
    train_dataset=train_dataset,
    eval_dataset=val_dataset,
    tokenizer=tokenizer_T5_tuned
)
trainer.train()

#save 
trainer.save_model("/kaggle/working/trainer")
tokenizer_T5_tuned.save_pretrained("/kaggle/working/tokenizer")

#load
tokenizer_T5_tuned = T5Tokenizer.from_pretrained("/kaggle/input/trained-nlp-models/tokenizer_t5/tokenizer")
model_T5_tuned = T5ForConditionalGeneration.from_pretrained("/kaggle/input/trained-nlp-models/model_t5/trainer")
model_T5_tuned.to(device)

### Evaluation: Fine-Tuned Model

#avg_bleu, avg_em, avg_f1=eval_answers(model_T5_tuned,tokenizer_T5_tuned,test_data,max_input=512,max_output=64)
print_scores(model_name="t5_tuned",use_backup=True)

### let's test the previus answer

tuned_model_answer=generate_answer(model_T5_tuned, tokenizer_T5_tuned, question=sample_zero_shot["question"].values[0], device=device, max_input_length=512, max_output_length=64, context=sample_zero_shot["context"].values[0])
sample_zero_shot["question"].values[0],tuned_model_answer, sample_zero_shot["answer"].values[0]

## 7. Hyperparameter Tuning on 15% of the Dataset

In order to optimize the T5 model performance, we performed a hyperparameter tuning process using only 15% of the training dataset. 

The following table summarizes the results of 27 different trials. Each trial is evaluated using a composite value score, which considers BLEU, EM, and F1. The best performing configurations are those with the highest `value`.

Hyperparameters tuned:
- **Learning Rate (lr)**: Controls how fast the model adapts during training.
- **Batch Size**: Affects stability and memory usage during training.
- **Epochs**: Number of times the model sees the entire dataset.
- **Learning Rate Scheduler (lr_sched)**: Strategy to adjust the learning rate over time.
- **Weight Decay**: Regularization term to prevent overfitting.

### Results 
Below is the table with trial configurations and their corresponding scores:

import optuna 
study = optuna.load_study(
        study_name='t5_small',
        storage=f'sqlite:////kaggle/input/nlp-resources/t5_small_hyper_parameter_tuning.db')
import pandas as pd
data = []
for trial in study.trials:
    trial_data = trial.params.copy()
    trial_data.update(trial.user_attrs)
    trial_data["value"] = trial.value  # score o loss principale
    trial_data["trial_number"] = trial.number
    data.append(trial_data)

trials_df = pd.DataFrame(data)[:27]
trials_df


## 8. Final Comparison of All Model Configurations

This table summarizes the performance of all major configurations evaluated throughout the project.

| Model Configuration         | Context Used | Training Type     |On all test data |Exact Match (%) | F1 Score (%) | BLEU Score |
|----------------------------|--------------|-------------------|-----------------|--------------|------------|------------|
| Zero-Shot                  | yes           | Pretrained only   | True            | 0.30            | 19.71         | 2.79       |
| Zero-Shot (No Context)     | no            | Pretrained only   | True            | 0.00            | 15.93         | 1.53       |
| Few-Shot (1 example)       | yes           | Prompting only    | False (1 sample)            | 0.00            | 62.0         | 31       |
| Fine-Tuned (Best Config)   | yes           | Supervised tuning | True            | 10.11           | 59.45         | 39.32       |
